+++ packages/vscode-extension/src/linterService.ts
@@ -1,9 +1,15 @@
-import { linterFileToServerVersion } from '@neo4j-cypher/lint-worker';
+import {
+  linterFileToServerVersion,
+  NpmRelease,
+} from '@neo4j-cypher/lint-worker';
 import { getExtensionContext } from './contextService';
 import * as vscode from 'vscode';
 import { sendNotificationToLanguageClient } from './languageClientService';
 import { linterStatusBarItem } from './extension';
 import { CONSTANTS } from './constants';
+import axios from 'axios';
+import * as tar from 'tar';
+import { pipeline } from 'stream/promises';
 
 export async function getFilesInExtensionStorage(): Promise<string[]> {
   const storageUri = await getStorageUri();
@@ -42,8 +48,89 @@ export async function switchWorkerOnLanguageServer(
   linterStatusBarItem.text = linterVersion ? linterVersion : 'Default';
 }
 
-async function getStorageUri(): Promise<vscode.Uri> {
+export async function getStorageUri(): Promise<vscode.Uri> {
   const storageUri = getExtensionContext().globalStorageUri;
   await vscode.workspace.fs.createDirectory(storageUri);
   return storageUri;
+}
+
+/**
+ * Attempts to download the latest linter of version $linterVersion,
+ * saving the file as "$linterVersion-lintWorker-$npmVersion.cjs"
+ * returning the name of the file saved if successful, and a blank string otherwise
+ */
+export async function downloadLintWorker(
+  linterVersion: string,
+  storageUri: vscode.Uri,
+  npmReleases: NpmRelease[],
+): Promise<boolean> {
+  void vscode.window.showInformationMessage(
+    `Downloading linter ${linterVersion} for your server`,
+  );
+
+  const newestLegacyLinter = npmReleases?.find(
+    (release) => release.tag === `neo4j-${linterVersion}`,
+  );
+  if (!newestLegacyLinter) {
+    void vscode.window.showErrorMessage(
+      CONSTANTS.MESSAGES.LINTER_VERSION_NOT_AVAILABLE,
+    );
+    return false;
+  }
+  const fileName = `${linterVersion}-lintWorker-${newestLegacyLinter.version}.cjs`;
+  const fileUri = vscode.Uri.joinPath(storageUri, fileName);
+
+  const downloadUrl = `https://registry.npmjs.org/@neo4j-cypher/lint-worker/-/lint-worker-${newestLegacyLinter.version}.tgz`;
+  try {
+    const response = await axios.get(downloadUrl, { responseType: 'stream' });
+    await pipeline(
+      response.data,
+      tar.x({
+        cwd: storageUri.fsPath,
+        filter: (path) => path === 'package/dist/cjs/lintWorker.cjs',
+      }),
+    );
+
+    const extractedUri = vscode.Uri.joinPath(
+      storageUri,
+      'package',
+      'dist',
+      'cjs',
+      'lintWorker.cjs',
+    );
+    const newFolderUri = vscode.Uri.joinPath(storageUri, 'package');
+    await vscode.workspace.fs.rename(extractedUri, fileUri, {
+      overwrite: true,
+    });
+    await vscode.workspace.fs.delete(newFolderUri, { recursive: true });
+    await deleteOutdatedLinters(linterVersion, fileName, storageUri);
+    return true;
+  } catch (error) {
+    void vscode.window.showErrorMessage(
+      CONSTANTS.MESSAGES.LINTER_DOWNLOAD_FAILED,
+    );
+    return false;
+  }
+}
+
+export async function deleteOutdatedLinters(
+  linterVersion: string,
+  newFile: string,
+  storageUri: vscode.Uri,
+) {
+  const fileNames = await getFilesInExtensionStorage();
+  const outDatedLinters: Record<string, string> = Object.fromEntries(
+    fileNames
+      .map((name) => [linterFileToServerVersion(name), name])
+      .filter(
+        (v): v is [string, string] =>
+          v !== undefined &&
+          v.length === 2 &&
+          v[0] === linterVersion &&
+          v[1] !== newFile,
+      ),
+  );
+  for (const [, fileName] of Object.entries(outDatedLinters)) {
+    await vscode.workspace.fs.delete(vscode.Uri.joinPath(storageUri, fileName));
+  }
 }
